automations/invoice-webapp/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pdfplumber==0.10.3
requests==2.31.0
gunicorn==21.2.0
